cJSON library: Improve precision for myfloor()
Fixes a bug where large MI tree integers would get printed as:
    {
      "shmem:total_size": 2.147484e+09,
      ...
    }

... due to lack of precision.

(cherry picked from commit abbb728f6b52bb3dd633ab3644709badf80986fe)

+++ lib/cJSON.c
@@ -69,8 +69,8 @@
 #else
 	double myfloor(double x)
 	{
-		if ( x > 0 ) return (int)x;
-		return (int)(x-0.9999999999999999);
+		if ( x > 0 ) return (long)x;
+		return (long)(x-0.9999999999999999);
 	}
 	#define FLOOR myfloor
 #endif
